requirements.txt
asyncpg
fastapi[all]
fastapi-users[beanie]
python-dotenv
motor
